+++ packages/flutter_deck/lib/src/renderers/renderers.dart
@@ -1,1 +1,1 @@
-export 'flutter_slide_image_renderer.dart';
+export 'flutter_deck_slide_image_renderer.dart';